+++ cmd/loomd/rlimit_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package main
+
+import "log/slog"
+
+func tuneNoFileLimit(logger *slog.Logger) {}